# onvif/api/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "接口文档",
        "contact": {},
        "version": "1.0.0"
    },
    "paths": {
        "/media_server_groups": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务分组"
                ],
                "summary": "列表",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "条数，小于 1 不匹配",
                        "name": "count",
                        "in": "query"
                    },
                    {
                        "maxLength": 128,
                        "type": "string",
                        "description": "描述，模糊",
                        "name": "describe",
                        "in": "query"
                    },
                    {
                        "maxLength": 32,
                        "type": "string",
                        "description": "名称，模糊",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "minimum": 0,
                        "type": "integer",
                        "description": "偏移，小于 0 不匹配",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "排序，\"column [desc]\"",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.GORMList-db_MediaServerGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务分组"
                ],
                "summary": "添加",
                "parameters": [
                    {
                        "description": "数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mediaservergroups.postReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal.IDResult-int64"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "delete": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务分组"
                ],
                "summary": "批量删除",
                "parameters": [
                    {
                        "description": "id 数组",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal.BatchDelete-int64"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            }
        },
        "/media_server_groups/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务分组"
                ],
                "summary": "详情",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/db.MediaServerGroup"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "流媒体服务分组"
                ],
                "summary": "删除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            },
            "patch": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "流媒体服务分组"
                ],
                "summary": "修改",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "数据",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mediaservergroups.patchReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal.RowResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "db.MediaServerGroup": {
            "type": "object",
            "properties": {
                "describe": {
                    "description": "描述",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "description": "名称",
                    "type": "string"
                }
            }
        },
        "internal.BatchDelete-int64": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "description": "数据库 ID 数组",
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
...
        "internal.Error": {
            "type": "object",
            "properties": {
                "detail": {
                    "description": "详细信息",
                    "type": "string"
                },
                "phrase": {
                    "description": "简短信息",
                    "type": "string"
                }
            }
        },
        "internal.IDResult-int64": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "数据库 ID",
                    "type": "integer"
                }
            }
        },
        "internal.RowResult": {
            "type": "object",
            "properties": {
                "row": {
                    "description": "行数",
                    "type": "integer"
                }
            }
        },
        "mediaservergroups.patchReq": {
            "type": "object",
            "properties": {
                "describe": {
                    "description": "描述",
                    "type": "string",
                    "maxLength": 128
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
        "mediaservergroups.postReq": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "describe": {
                    "description": "描述",
                    "type": "string",
                    "maxLength": 128
                },
                "name": {
                    "description": "名称",
                    "type": "string",
                    "maxLength": 32
                }
            }
        },
...
        "util.GORMList-db_MediaServerGroup": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/db.MediaServerGroup"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        }
    }
}